+++ tests/__init__.py
@@ -1,1 +1,1 @@
-from . import test_instructor_school, test_scaffold
+from . import test_course_enrollment, test_instructor, test_instructor_school, test_scaffold